refactor: move globalStyles into stitches.config
