+++ site/src/styles/index.ts
@@ -1,4 +1,3 @@
 export * from './components'
-export * from './global'
 export * from './Prose'
 export * from './themes'
